\name{NDVI.Example1}
\alias{NDVI.Example1}
\encoding{UTF-8}
\docType{data}
\title{
Site 1
}
\description{
NDVI Test Site 1 (Jamali et al. 2015)
}
\usage{data("NDVI.Example1")}
\format{
  The format is:
 num [1:300] 0.507 0.582 0.466 0.661 0.708 ...
}
\source{
Jamali, S, Eklundh, L, Ardö, J, Seaquist, J, & Jönsson, P 2015, 'Detecting changes in vegetation trends using
time series segmentation', Remote Sensing Of Environment, 156, p. 182-195, Scopus®, EBSCOhost,
viewed 27 March 2015.
}
\examples{
data(NDVI.Example1)
## maybe str(NDVI.Example1) ; plot(NDVI.Example1) ...
}
\keyword{datasets}
