\name{NDVI.Example2}
\alias{NDVI.Example2}
\encoding{UTF-8}
\docType{data}
\title{
Site 2
}
\description{
NDVI Test Site 2 (Jamali et al. 2015)
}
\usage{data("NDVI.Example2")}
\format{
  The format is:
 num [1:300] 0.125 0.2 0.121 0.155 0.206 ...
}

\source{
Jamali, S, Eklundh, L, Ardö, J, Seaquist, J, & Jönsson, P 2015, 'Detecting changes in vegetation trends using
time series segmentation', Remote Sensing Of Environment, 156, p. 182-195, Scopus®, EBSCOhost,
viewed 27 March 2015.
}

\examples{
data(NDVI.Example2)
## maybe str(NDVI.Example2) ; plot(NDVI.Example2) ...
}
\keyword{datasets}
