\name{NDVI.Site1}
\alias{NDVI.Site1}
\encoding{UTF-8}
\docType{data}
\title{
Site 1
}
\description{
NDVI Test data (Site 1 in Jamali et al. 2015)
}
\usage{data("NDVI.Site1")}
\format{
  The format is:
 num [1:300] 0.507 0.582 0.466 0.661 0.708 ...
}
\source{
Jamali S, Jönsson P, Eklundh L, Ardö J, Seaquist J (2015). Detecting changes in vegetation trends using time series segmentation. Remote Sensing of Environment, 156, 182-195. http://dx.doi.org/10.1016/j.rse.2014.09.010
}
\examples{
data(NDVI.Site1)
## maybe str(NDVI.Site1) ; plot(NDVI.Site1) ...
}
\keyword{datasets}
