\name{NDVI.Site2}
\alias{NDVI.Site2}
\encoding{UTF-8}
\docType{data}
\title{
Site 2
}
\description{
NDVI Test data (Site 2 in Jamali et al. 2015)
}
\usage{data("NDVI.Site2")}
\format{
  The format is:
 num [1:300] 0.125 0.2 0.121 0.155 0.206 ...
}

\source{
Jamali S, Jönsson P, Eklundh L, Ardö J, Seaquist J (2015). Detecting changes in vegetation trends using time series segmentation. Remote Sensing of Environment, 156, 182-195. http://dx.doi.org/10.1016/j.rse.2014.09.010
}

\examples{
data(NDVI.Site2)
## maybe str(NDVI.Site2) ; plot(NDVI.Site2) ...
}
\keyword{datasets}
